WIP: Repariere Import-Pfade nach Service-Reorganisation
Teilweise behobene Imports:
- dependencies.py: Alle Service-Imports aktualisiert
- app/services/__init__.py: Service-Pfade korrigiert
- main.py: Migration, logging, cleanup, startup_init Pfade gefixt
- middleware/auth.py: auth_service Pfad korrigiert
- routes/auth.py: auth_service Pfad korrigiert
- routes/recording.py: logging_service Pfad korrigiert
- events/handler_registry.py: websocket_manager Pfad korrigiert

Communication Services:
- __init__.py: Klassen und Instances exportiert
- enhanced_push_service: Import-Pfad korrigiert
- webpush_service: Instance hinzugefügt (None wegen missing args)
- websocket_manager: Instance hinzugefügt

Core Services:
- __init__.py: Klassen statt Instances exportiert

NOCH ZU TUN: Viele weitere Import-Pfade in allen Routes und Services

Amp-Thread: https://ampcode.com/threads/T-54f91e8c-d69a-4ab5-b57a-d07402749a17
