+++ app/services/communication/__init__.py
@@ -4,12 +4,12 @@ Communication Services Package
 Contains notification and communication services.
 """
 
-from .webpush_service import webpush_service
-from .enhanced_push_service import enhanced_push_service
-from .websocket_manager import websocket_manager
+from .webpush_service import ModernWebPushService, webpush_service
+from .enhanced_push_service import EnhancedPushService, enhanced_push_service  
+from .websocket_manager import ConnectionManager, websocket_manager
 
 __all__ = [
-    'webpush_service',
-    'enhanced_push_service', 
-    'websocket_manager'
+    'ModernWebPushService', 'webpush_service',
+    'EnhancedPushService', 'enhanced_push_service', 
+    'ConnectionManager', 'websocket_manager'
 ]
