+++ app/services/core/__init__.py
@@ -4,14 +4,14 @@ Core Services Package
 Contains core application services.
 """
 
-from .auth_service import auth_service
-from .settings_service import settings_service
-from .state_persistence_service import state_persistence_service
-from .test_service import test_service
+from .auth_service import AuthService
+from .settings_service import SettingsService  
+from .state_persistence_service import StatePersistenceService, state_persistence_service
+from .test_service import StreamVaultTestService
 
 __all__ = [
-    'auth_service',
-    'settings_service',
-    'state_persistence_service',
-    'test_service'
+    'AuthService',
+    'SettingsService',
+    'StatePersistenceService', 'state_persistence_service',
+    'StreamVaultTestService'
 ]
