+++ app/services/media/thumbnail_service.py
